add test to project

+++ routes/admin.php
@@ -9,40 +9,40 @@ use App\Http\Controllers\admin\RoleController;
 use Illuminate\Support\Facades\Route;
 
 
-Route::delete('posts/{post}',[PostController::class,'destroy'])->middleware('auth:sanctum')->middleware('permission:delete_post');
-Route::delete('comment/{id}', [CommentController::class,'destroy'])->middleware('auth:sanctum')->middleware('permission:delete_comment');
-
+Route::delete('posts/{post}', [PostController::class, 'destroy'])
+    ->middleware('auth:sanctum')->middleware('permission:delete_post')->name('posts-deletePost');
+Route::delete('comment/{id}', [CommentController::class, 'destroy'])
+    ->middleware('auth:sanctum')->middleware('permission:delete_comment')->name('comments-deleteComment');
 
 
 Route::middleware('auth:sanctum')->group(function () {
 
-    Route::get('/pending_posts',[PostController::class,'index'])->middleware('permission:show_post');
-    Route::get('/post/{post}',[PostController::class,'show'])->middleware('permission:show_post');
-    Route::patch('/confirm_post/{post}',[PostController::class,'update'])->middleware('permission:confirm_post');
+    Route::get('/pending_posts', [PostController::class, 'index'])->middleware('permission:show_post')->name('posts-pending');
+    Route::get('/post/{post}', [PostController::class, 'show'])->middleware('permission:show_post')->name('posts-status');
+    Route::patch('/confirm-post/{post}', [PostController::class, 'update'])->middleware('permission:confirm_post')->name('posts-confirm');
 });
 
 
 Route::prefix('categories')->middleware('auth:sanctum')->group(function () {
 
-    Route::get('/', [CategoryController::class, 'index']);
-    Route::post('/', [CategoryController::class, 'store'])->middleware('permission:add_category');
-    Route::delete('/{category}', [CategoryController::class, 'destroy'])->middleware('permission:delete_category');
-    Route::patch('/{category}/update', [CategoryController::class, 'update'])->middleware('permission:update_category');
+    Route::get('/', [CategoryController::class, 'index'])->name('categories-all');
+    Route::post('/', [CategoryController::class, 'store'])->middleware('permission:add_category')->name('categories-add');
+    Route::delete('/{category}', [CategoryController::class, 'destroy'])->middleware('permission:delete_category')->name('categories-delete');
+    Route::patch('/{category}/update', [CategoryController::class, 'update'])->middleware('permission:update_category')->name('categories-update');
 });
-
 
 
 Route::prefix('super_admin')->middleware(['auth:sanctum', 'role:super_admin'])->group(function () {
 
-    Route::get('roles',[RoleController::class,'index']);
-    Route::get('roles/{role}',[RoleController::class,'show']);
-    Route::post('roles', [RoleController::class, 'store']);        //add role
-    Route::patch('roles/{role}', [RoleController::class,'update']);
-    Route::delete('roles/{role}',[RoleController::class,'delete']);
+    Route::get('roles', [RoleController::class, 'index'])->name('roles-all');
+    Route::get('roles/{role}', [RoleController::class, 'show'])->name('roles-oneRole');
+    Route::post('roles', [RoleController::class, 'store'])->name('roles-addRole');        //add role
+    Route::patch('roles/{role}', [RoleController::class, 'update'])->name('roles-updateRole');
+    Route::delete('roles/{role}', [RoleController::class, 'delete'])->name('roles-deleteRole');
 
-    Route::get('permissions',[PermissionController::class,'index']);
+    Route::get('permissions', [PermissionController::class, 'index'])->name('permissions-all');
 
-    Route::post('admins', [AdminController::class, 'store']);                   //create admin
-    Route::patch('admins/user/{user}/role',[AdminController::class,'update']);  //add role to exist user
-    Route::delete('admins/user/{user}/role',[AdminController::class,'destroy']);//remove role from admins
+    Route::post('admins', [AdminController::class, 'store'])->name('admins-add');                      //create admin
+    Route::patch('admins/user/{user}/role', [AdminController::class, 'update'])->name('admins-update');  //add role to existing user
+    Route::delete('admins/user/{user}/role', [AdminController::class, 'destroy'])->name('admins-delete');//remove role from admins
 });
